feat: fuzzy filtering via 'completefilterfunc'
Introduces a new option 'completefilterfunc', to filter insert
completion matches. completefilterfunc must be a function with
signature

  function(prefix, match)

and return a score in the interval [0, 999999). Higher scoring
matches are placed higher up on the completion popup menu, and
a score of zero removes the match from being shown.
'completefilterfunc' is called when starting insert mode completion,
and whenever the user types a character to narrow down the
completion list, allowing for native fuzzy filtering.

This PR also introduces a new api function 'nvim_complete',
which allows for overriding the 'completefilterfunc' option
with another filterfunc for this completion only.